FixedActivities.c- changed LongPowerFailTimer from uint8_t to uint_16 to support a count upto 600 -Changed energy accumulation routine -Added 32 by 32 bit multiplier routine -Added a callback function for meter calibrating parameters.

+++ emeter-ng/SM502/FixedActivities.c
@@ -33,7 +33,7 @@ extern uint8_t CoverRemovalTripDuration    ;
  uint8_t UnderVoltageTimer  ;
  uint8_t UnderFrequencyTimer;
  uint8_t OverFrequencyTimer ;
- uint8_t LongPowerFailTimer ;
+ uint16_t LongPowerFailTimer ;
  
 extern  MeterStatus status; 
 volatile uint8_t secondCounter =0;
